requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
request s==2.31.0
